backend/requirements.txt
#Web Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23

# RAG Components
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.3.0
chromadb>=0.4.22

# LLM and ML
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
sentencepiece>=0.1.99

# For GPU support (optional, comment out if no GPU)
# bitsandbytes>=0.41.3

# Image Processing - Use latest compatible version
Pillow>=10.2.0
pytesseract>=0.3.10

# Security and Auth
bcrypt>=4.1.2
PyJWT>=2.8.0
bleach>=6.1.0
werkzeug>=3.0.1

# Utilities
python-multipart>=0.0.6
numpy>=1.26.0
scikit-learn>=1.3.0
tqdm>=4.66.0